# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlyncPay API",
    "version": "2.0.0",
...
  },
  "servers": [
    { "url": "https://slyncpay-api.onrender.com", "description": "Production" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "validation_error" },
          "message": { "type": "string", "example": "Field xyz is required" },
          "statusCode": { "type": "integer", "example": 400 }
        }
      },
...
      "LineItem": {
        "type": "object",
        "required": ["amountCents"],
        "properties": {
          "description": { "type": "string", "maxLength": 500 },
          "amountCents": { "type": "integer", "minimum": 1 },
          "quantity": { "type": "number", "minimum": 0 },
          "unit": { "type": "string", "maxLength": 50 }
        }
      },
      "Payable": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "entityId": { "type": "string", "format": "uuid" },
          "workerId": { "type": "string", "format": "uuid" },
          "engagementId": { "type": "string", "format": "uuid" },
          "externalReferenceId": { "type": ["string", "null"], "description": "Your reference (e.g. invoice or shift ID)." },
          "amountCents": { "type": "integer", "minimum": 1 },
          "status": { "type": "string", "enum": ["draft", "pending", "processing", "paid", "failed", "cancelled"] },
          "dueDate": { "type": "string", "format": "date" },
          "lineItems": { "type": "array", "items": { "$ref": "#/components/schemas/LineItem" } },
          "disbursementId": { "type": ["string", "null"], "format": "uuid", "description": "Set once the payable has been swept into a disbursement batch." },
          "paidAt": { "type": ["string", "null"], "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/v1/payables": {
      "post": {
        "tags": ["Payables"],
        "summary": "Create a payable (queued for next disbursement)",
        "description": "Creates a pending payable for the given worker + entity. It will be paid the next time you trigger a disbursement for that entity.\n\nContractor must have `onboardingStatus = active`; otherwise returns 400 (`validation_error`).\n\n**Idempotency-Key header required.**",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["entityId", "workerId", "amountCents", "dueDate", "lineItems"],
                "properties": {
                  "entityId": { "type": "string", "format": "uuid" },
                  "workerId": { "type": "string", "format": "uuid" },
                  "externalReferenceId": { "type": "string", "maxLength": 200 },
                  "amountCents": { "type": "integer", "minimum": 1 },
                  "dueDate": { "type": "string", "format": "date" },
                  "lineItems": { "type": "array", "items": { "$ref": "#/components/schemas/LineItem" }, "minItems": 1 },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              },
              "example": {
                "entityId": "9e1d…-…",
                "workerId": "7c3f…-…",
                "externalReferenceId": "SHIFT-9021",
                "amountCents": 45000,
                "dueDate": "2026-05-15",
                "lineItems": [{ "description": "Shift on 2026-05-12", "amountCents": 45000 }]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Payable" } } } },
          "400": { "description": "Worker not active, no engagement, or other validation failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["Payables"],
        "summary": "List payables",
        "parameters": [
          { "name": "entityId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "workerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "pending", "processing", "paid", "failed", "cancelled"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Payable" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/payables/{id}": {
      "get": {
        "tags": ["Payables"],
        "summary": "Get a payable",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Payable" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}